Integrate critic optimizer in td3 and remove dpg (#81)
* Merge: integrate critic optimizer in td3

Squashed commit of the following:

commit a57e58f053c6858f3fee7e900068677dd263832d
Author: khkim <kh.kim@medipixel.io>
Date:   Thu Feb 14 14:42:27 2019 +0900

    Modify integrating critic on td3 in lunarlander

commit 1edceed6feb66e661e395746496f0bafce98fe81
Author: khkim <kh.kim@medipixel.io>
Date:   Thu Feb 14 10:33:00 2019 +0900

    Test integrating critic optimizer in td3

commit 3e27403814e36d449a0841ed272764eaeb254e7d
Author: khkim <kh.kim@medipixel.io>
Date:   Thu Feb 14 10:10:00 2019 +0900

    Move increasing step in td3

commit 7302d77d750aefbbf4ef493e77bf1d98b8b63902
Author: khkim <kh.kim@medipixel.io>
Date:   Thu Feb 14 09:58:09 2019 +0900

    Remove unused agent

* Remove dpg
