+++ README.md
@@ -6,11 +6,10 @@ RL key algorithms with [LunarLanderContinuous-v2](https://gym.openai.com/envs/Lu
 
 1. Reinforce with baseline
 2. Advantage Actor-Critic (A2C)
-3. Deterministic Policy Gradient (DPG)
-4. Deep Deterministic Policy Gradient (DDPG)
-5. Proximal Policy Optimization Algorithms (PPO)
-6. Twin Delayed Deep Deterministic Policy Gradient Algorithm (TD3)
-7. Soft Actor Critic Algorithm (SAC)
-8. Behaviour Cloning (BC with DDPG)
-9. Prioritized Experience Replay (PER with DDPG, TD3, SAC)
-10. From Demonstrations (DDPGfD, TD3fD, SACfD)
+3. Deep Deterministic Policy Gradient (DDPG)
+4. Proximal Policy Optimization Algorithms (PPO)
+5. Twin Delayed Deep Deterministic Policy Gradient Algorithm (TD3)
+6. Soft Actor Critic Algorithm (SAC)
+7. Behaviour Cloning (BC with DDPG)
+8. Prioritized Experience Replay (PER with DDPG, TD3, SAC)
+9. From Demonstrations (DDPGfD, TD3fD, SACfD)
