--- examples/lunarlander_continuous_v2/dpg.py
@@ -1,74 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Run module for DPG on LunarLanderContinuous-v2.
-
-- Author: Curt Park
-- Contact: curt.park@medipixel.io
-"""
-
-import argparse
-
-import gym
-import torch
-import torch.optim as optim
-
-from algorithms.common.networks.mlp import MLP
-from algorithms.dpg.agent import Agent
-
-device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-
-# hyper parameters
-hyper_params = {
-    "GAMMA": 0.99,
-    "LR_ACTOR": 1e-3,
-    "LR_CRITIC": 1e-3,
-    "WEIGHT_DECAY": 1e-6,
-}
-
-
-def run(env: gym.Env, args: argparse.Namespace, state_dim: int, action_dim: int):
-    """Run training or test.
-
-    Args:
-        env (gym.Env): openAI Gym environment with continuous action space
-        args (argparse.Namespace): arguments including training settings
-        state_dim (int): dimension of states
-        action_dim (int): dimension of actions
-
-    """
-    # create models
-    actor = MLP(
-        input_size=state_dim,
-        output_size=action_dim,
-        hidden_sizes=[128, 128],
-        output_activation=torch.tanh,
-    ).to(device)
-
-    critic = MLP(
-        input_size=state_dim + action_dim, output_size=1, hidden_sizes=[256, 256]
-    ).to(device)
-
-    # create optimizer
-    actor_optim = optim.Adam(
-        actor.parameters(),
-        lr=hyper_params["LR_ACTOR"],
-        weight_decay=hyper_params["WEIGHT_DECAY"],
-    )
-
-    critic_optim = optim.Adam(
-        critic.parameters(),
-        lr=hyper_params["LR_CRITIC"],
-        weight_decay=hyper_params["WEIGHT_DECAY"],
-    )
-
-    # make tuples to create an agent
-    models = (actor, critic)
-    optims = (actor_optim, critic_optim)
-
-    # create an agent
-    agent = Agent(env, args, hyper_params, models, optims)
-
-    # run
-    if args.test:
-        agent.test()
-    else:
-        agent.train()
